created bitcoinServer.js and yelpServer.js, updated naming conventions
