+++ server/bitcoinServer.js
@@ -3,27 +3,22 @@ const runStream = require('./stream');
 const WebSocket = require('ws');
 const fs = require('fs');
 const kafka = require('./kafka/kafkaConnection');
-// const createConsumer = require('./kafka/streams/transactionConsumer');
-const createConsumer = require('./kafka/bitcoin/transactionConsumer');
+const createConsumer1 = require('./kafka/bitcoin/transactionConsumer');
+// const createConsumer2 = require('./kafka/bitcoin/calculatedTransactionConsumer');
 // const path = './myapp.log';
 
-// const connectProducer = async () => {
 const producer = kafka.producer();
 producer.connect();
 
-createConsumer(kafka, 'transactions');
+createConsumer1(kafka, 'transactions');
 // createConsumer2(kafka, 'calculatedTransactions');
 
-//   return producer;
-// };
-// const producer = connectProducer();
-
 async function sendEvent(data, topic) {
-  const senderAddress = data.x.inputs[0].prev_out.addr;
-  console.log('sender address is', senderAddress);
+  //   const senderAddress = data.x.inputs[0].prev_out.addr;
+  //   console.log('sender address is', senderAddress);
   await producer.send({
     topic,
-    messages: [{ value: JSON.stringify(senderAddress) }],
+    messages: [{ value: JSON.stringify(data) }],
   });
 }
 
@@ -33,7 +28,6 @@ fs.writeFile('myapp.log', '', (err) => {
 });
 
 const app = express();
-// runStream();
 
 const ws = new WebSocket('wss://ws.blockchain.info/inv', {
   perMessageDeflate: false,
@@ -46,13 +40,15 @@ ws.on('open', () => {
 
 ws.on('message', (data) => {
   const transaction = JSON.parse(data);
-  // console.log(data);
-  const amount = transaction.x.inputs[0].prev_out.value / 100000000;
+  // console.log(JSON.stringify(data, null, 2));
+  // console.log(JSON.parse(transaction.x.inputs[0].prev_out.value, null, 2));
+  const amount = JSON.parse(transaction.x.inputs[0].prev_out.value) / 100000000;
+  console.log(amount);
   const dollars = amount * 19122;
   if (amount) {
-    sendEvent(transaction, 'transactions');
-    console.log(`Amount: ${amount}`);
-    console.log(`Dolla dolla bill: ${dollars}`);
+    sendEvent(data, 'transactions');
+    // console.log(`Amount: ${amount}`);
+    // console.log(`Dolla dolla bill: ${dollars}`);
     // console.log(`Size: ${transaction.x.size}`);
     // console.log(`In Addr: ${transaction.x.inputs[0].prev_out.addr}`);
     // console.log(`Out addr: ${transaction.x.out[0].addr}`);
